backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
nltk
sqlalchemy
python-jose[cryptography]